Combat, TurnCountUI and GameOverUI
minor fix to combat so that after being initiated it checks if any player's health is less than or equal to 0, at which point their status is switched to dead.

Turn Count UI is located in top center of the canvas.

Game Over UI pops up at the very end when the gameOver bool is set to true.